refactor: rename util files
